fix: add special_tx file tests
